Moved many HistoryBank queries to AbstractHistory.
