+++ app/src/main/java/pen/history/History.java
@@ -7,6 +7,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.function.Function;
+import java.util.function.Predicate;
 
 /**
  * A {@link HistoryBank} or {@link HistoryQuery}.
@@ -30,6 +31,7 @@ public interface History {
     List<Incident> getIncidents();
 
     TimeFrame getTimeFrame();
+    TimeFrame getTimeFrame(Predicate<Incident> filter);
 
     Map<String,Period> getPeriods(TimeFrame frame);
