+++ app/src/main/java/pen/history/HistoryBank.java
@@ -45,101 +45,6 @@ public class HistoryBank
         return incidents();
     }
 
-    /**
-     * Returns the time frame given all events.
-     * @return the time frame.
-     */
-    public TimeFrame getTimeFrame() {
-        return getTimeFrame(incident -> true);
-    }
-
-    public TimeFrame getTimeFrame(Predicate<Incident> filter) {
-        var filtered = incidents().stream()
-            .filter(filter)
-            .toList();
-
-        var start = filtered.stream()
-            .mapToInt(Incident::moment)
-            .min().orElse(0);
-        var end = filtered.stream()
-            .mapToInt(Incident::moment)
-            .max().orElse(0);
-
-        return new TimeFrame(start, end);
-    }
-
-    public Optional<Period> getPeriod(String entityId) {
-        return getPeriod(entityId, getTimeFrame());
-    }
-
-    public Optional<Period> getPeriod(String entityId, TimeFrame frame) {
-        // FIRST, get the entity
-        var entity = entityMap().get(entityId);
-        if (entity == null) {
-            throw new IllegalArgumentException("No such entity: \"" + entityId + "\"");
-        }
-
-        // NEXT, get the sorted earliest and last incidents concerning this
-        // entity
-        var all = getIncidents(entityId);
-
-        if (all.isEmpty()) {
-            return Optional.empty();
-        }
-
-        var first = all.getFirst();
-        var last = all.getLast();
-
-        int startMoment;
-        int endMoment;
-        Cap startCap;
-        Cap endCap;
-
-        if (first.moment() >= frame.end()) {
-            return Optional.empty();
-        } else if (first.moment() < frame.start()) {
-            startMoment = frame.start();
-            startCap = Cap.SOFT;
-        } else {
-            startMoment = first.moment();
-            startCap = first.cap();
-        }
-
-        if (last.moment() <= frame.start()) {
-            return Optional.empty();
-        } else if (last.moment() > frame.end()) {
-            endMoment = frame.end();
-            endCap = Cap.SOFT;
-        } else {
-            endMoment = last.moment();
-            endCap = last.cap();
-        }
-
-        return Optional.of(new Period(
-            entity,
-            startMoment,
-            endMoment,
-            startCap,
-            endCap
-        ));
-    }
-
-    public Map<String,Period> getPeriods(TimeFrame frame) {
-        var map = new HashMap<String,Period>();
-
-        for (var id : entityMap().keySet()) {
-            getPeriod(id, frame).ifPresent(period -> map.put(id, period));
-        }
-
-        return map;
-    }
-
-    public List<Incident> getIncidents(String entityId) {
-        return incidents().stream()
-            .filter(i -> i.concerns(entityId))
-            .sorted(Comparator.comparing(Incident::moment))
-            .toList();
-    }
 
     private static final String INCIDENTS = "Incidents";
     private static final String H_LINE = TextCanvas.LIGHT_HORIZONTAL;
